+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [7.39.1] - 2019-04-02
+### Changed
+- Make the query string part of the `href` in Render's Link component.
+
 ## [7.39.0] - 2019-03-13
 
 ## [7.38.13] - 2019-03-08